# rt: reduce no-op wakeups in the multi-threaded scheduler