@carllerche 2022-01-12
@Darksonn @hawkw this should be ready for review, it would be nice to have it confirmed in a "real app" too.