@blt 2022-01-12
> @Darksonn @hawkw this should be ready for review, it would be nice to have it confirmed in a "real app" too.

I'm planning on wiring this up into [vector](https://github.com/vectordotdev/vector) and running it through our integrated benchmarks. Assuming all goes well I'll link numbers here in a handful of hours.