@blt 2022-01-12
No change in observed vector throughput, per [these results](https://github.com/vectordotdev/vector/pull/10823#issuecomment-1011424094). That said, we don't track CPU use in our experimental setup and the observation duration in a PR is constrained to 200 seconds to make turn-around time feasible.