@blt 2022-01-12
Unfortunately we don't capture that kind of saturation information yet, though we do have an issue for it: https://github.com/vectordotdev/vector/issues/10456. That said, we're intending to fully saturate vector so it's reasonable to believe there's little idle time.